src/requirements.txt
absl-py==0.7.1
astor==0.8.0
awscli==1.14.58
certifi==2019.6.16
cort==0.2.4.5
docutils==0.14
future==0.16.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.7.1
html5lib==0.9999999
Jinja2==2.10
JPype1==0.6.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.0
mmh3==2.5.1
msgpack-python==0.5.4
numpy==1.17.0
Pillow==6.1.0
protobuf==3.9.1
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
pyhocon==0.3.51
pyparsing==2.2.0
PyStanfordDependencies==0.3.1
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
typing==3.6.4
Werkzeug==0.14.1
wrapt==1.11.2
keras-self-attention==0.42.0
